refactoring out static arrays/objects
