+++ src/parser.ts
@@ -2,10 +2,9 @@ import { getDomain } from 'tldts';
 import { getCanvasFontFp, getCanvasFp } from './canvas-fingerprinting';
 import { loadBrowserCookies, matchCookiesToEvents } from './cookie-collector';
 import { FB_ADVANCED_MATCHING_PARAMETERS, FB_STANDARD_EVENTS } from './fb-pixel-lookup';
+import { BEHAVIOUR_TRACKING_EVENTS, FINGERPRINTABLE_WINDOW_APIS } from './helpers/statics';
 import {
-    BEHAVIOUR_TRACKING_EVENTS,
     BlacklightEvent,
-    FINGERPRINTABLE_WINDOW_APIS,
     JsInstrumentEvent,
     KeyLoggingEvent,
     SessionRecordingEvent,
